Missing translation for Save Bookmark in Browser
When there are multiple Browsers installed, the dialog
"Complete action using" is displayed after clicking on
a star in Bookmarks screen. The activity AddBookmarkPage
has a hard-coded label name in English, which is displayed
in the dialog even in other languages. This change makes
it possible for developers to change the label name by
adding corresponding language resources.

Change-Id: Ibad4aa86ffc3467f2b67083827a9b256fdc4321a